Multiple worker backends on the web (to fix the Chrome crash) (#28)
Load WASM in an iframe instead of a Webworker to work around a Google Chrome out of memory error.

This commit enables defering the WASM workload to a configurable backend. It ships with three backends: `iframeWorkerBackend`, `webWorkerBackend`, `sharedWorkerBackend`. The `iframeWorkerBackend` is the default.

See #1 for more details about the Google Chrome crash.

**Important!** The iframe must be loaded from another domain to spin a new browser thread. If it's loaded from the same domain, WordPress "server" will run in the same thread that paints the user interface and dramatically slow down all user interactions.