Add loading progress bar (#46)
Loading WASM WordPress requires downloading about 20MB of files. Waiting without any feedback is a bad experience, not to mention you can never be sure whether it crashed or is still loading.

This commit introduces a progress bar. The idea is to capture the progress events in `wasm-worker.js`, pass them on to the main thread, and then use the information to update a CSS progress bar.

We can’t simply update the WASM loading code to capture the progress events because that codeis generated by emscripten on each build. Instead, this commit introduces a`WASMDownloadMonitor` class that:

* Monkeypatches WebAssembly.instantiateStreaming to capture the `php-wasm.wasm` progress events
* Observes emscripten’s `dataFileDownloads` using ES Proxies to capture the `wp.data` progress events

It also introduces a new worker backend method called `addMessageListener`. It’s necessary to propagate the progress information from the worker to the main thread.

Finally, it reduces the `init()` method to a simple wrapper that can be configured as needed in the app. This allows the seamless `wordpress.html` to just load WordPress and be done, and the more involved `wordpress-browser.html` to handle the progress bar internally.