PHPServer: Correctly resolve the request path to the file path
Unprefix the request path before resolving a file to serve. For example, for a server running at `http://localhost/root` and a request to `/root/index.css`, this commit makes sure that the filesystem path always gets resolved as `/index.css`.

Before this commit, PHPServer sometimes assumed the `/scope:19830` part of the request path is actually a filesystem path.