Pair the site editor's nested iframe to the Service Worker.
The site editor was initiating network requests that weren't routed
through the service worker. That's a known bug in Chrome, Firefox, and a
few other browsers based on these two. The issue is only with the
iframes using srcDoc and src="about:blank" as they fail to inherit the
root site's service worker.

Gutenberg loads the site editor using <iframe srcDoc="<!doctype html">
to force the standards mode and not the quirks mode:

https://github.com/WordPress/gutenberg/pull/38855

This commit patches the site editor to achieve the same result via
<iframe src="/doctype.html"> and a doctype.html file containing just
`<!doctype html>`. This allows the iframe to inherit the service worker
and correctly load all the css, js, fonts, images, and other assets.

Ideally this issue would be fixed directly in Gutenberg and the patch
below would be removed.

See https://github.com/WordPress/wordpress-wasm/issues/42 for more details