Refresh the service worker whenever a new version is available
Service workers are normally aggressively cached by the browsers. When a new version of the service worker is available, the browser may still cling to the old version for days. This commit forces the browser to refresh the service worker if a new version is available.

Technically:

* ESBuild prepares a preliminary build of the service worker script and computes the hash
* The hash is written to a service-worker-version.ts module
* Service worker imports the hash and exposes on the `/version` endpoint
* Frontend script imports the hash and compares it to the one received from `/version`
* If the two hashes don’t match, the frontend script force-refreshes the service worker

The preliminary build is always prepared with an empty version string to avoid an infinite loop of building, computing the hash, building a new version with the updated hash, computing a new hash etc.