Update the parameter used for URL paths
The parameter previously documented for the URL path was `page=/wp-admin/` but that doesn't work. Elsewhere in the documentation it mentions `&url=....` which got me thinking perhaps it was not named correctly in the current documentation. That led me to confirming that indeed `url` is the correct parameter to use here.