Bundle playground client dependencies
This is useful in itself and makes the client package self-contained. In
this instance we need it to avoid producing an invalid list of
dependencies in the final package.json. This package technically depends
on php-wasm/web, but takes great care not to import anything from it
other than TypeScript types that are then released in a rollup file.