Add Lerna (#184)
## Description

It's hard to publish all the npm packages living in this repo without a tool. After considering [changesets](https://github.com/changesets/changesets) and [Lerna](https://lerna.js.org/) this commit moves forward the latter since it's maintained by the authors of nx – another tool used in this project.

Solves https://github.com/WordPress/wordpress-playground/issues/148

This commit also bundles wp-playground/client dependencies

This is useful in itself and makes the client package self-contained. In
this instance we need it to avoid producing an invalid list of
dependencies in the final package.json. This package technically depends
on php-wasm/web, but takes great care not to import anything from it
other than TypeScript types that are then released in a rollup file.