Import/export support: WXR, WXZ, and full-site
This commit adds multiple ways to move data in and out of Playground.
Specifically, it implements the following client functions:

* zipEntireSite()
* replaceSite()
* exportWXR()
* exportWXZ()
* submitImporterForm()

The WXR format is one natively used by the official WordPress importer plugin.

The WXZ format is supported through the following plugin:

https://github.com/akirk/export-wxz
https://github.com/akirk/wordpress-importer

The full-site export is simply a zip archive containing the entire site.