Rename php-wasm/common to php-wasm/abstract and add php-wasm/util
The common package isn't intended for usage outside of php-wasm/node and
php-wasm/web, but the name "common" doesn't make it clear enough. This
commit introduces a better name.

In addition, there was no good place to
store utility classes and functions like Semaphore, which resulted in
re-exporting them from the web and node packages. This commit introduces
a `util` package as a new home for them.