Fix the truncated download of WordPress zip bundle
As seen in https://github.com/WordPress/wordpress-playground/issues/248,
the PHP files extracted from the WordPress zip bundle are truncated.
The root cause is an issue with unzipper.Extract(), possibly
interplaying with a deprecated `request` package.

This commit removes a `request` dependency and rewrites the unzipping
process using `unzipper.Parse()`