WP Now: Encapsulate logic related to different execution modes into distinct functions
I'm getting dizzy with all the if(mode === something) conditional in
different code paths. It's hard to reason about the logic executed for
any specific mode with so many forks.

This PR proposes replacing the current abstraction with four distinct functions:

* initInWpContentMode
* initInCoreDevelopMode
* initInCoreMode
* initInPluginOrThemeMode

They are narrowly focused on initializing that specific mode and have
almost no forks. It's very easy to read and make changes – you won't be
surprised by a nasty side-effect in another mode.

Also, turns out with this abstraction we don't even need the WPNow
class. It merely parsed the options and initialized PHP, we don't need a
shared state for that.