New documentation site built with Docusaurus
The current documentation suffers from a few problems:

* It is incomplete
* It is not interactive
* It is not connected to code and a PR can make it go stale
* It is not testable and there's no way to know whether the code
  snippets still work
* API docs can be hard to read
* Dead links – no link validation

This commit is the first step towards fixing these problems. It:

* Adds the missing bits of the documentation and improve the structure
* Adds the first interactive examples with more to come
* Generates parts of the docs directly from the doc strings
* Adds link validation
* Improves readability and interconnectedness between the content pages
  and the API reference

With this PR, adding a Blueprint step into the codebase will
automatically add it to the docs. If it has a usage example, it will
be rendered as a live code snippet. In the future, it will be editable
and the CI will verify whether it still works.