Playground website: Cache busting for remote.html
Loads remote.html from a build-dependent URL to ensure cache isn't used
for loading updated releases.

I noticed new deployments to playground.wordpress.net tend to break in Chrome.
The root cause is Chrome caches remote.html because the <iframe />
refers to the exact same URL before and after the deployment. As a
result, Chrome tries to load the old version of JS and wasm assets that
are no longer there and the cached ones are not compatible with the
refreshed website.