Meta: Remove the automated formatting check
The automated formatting check requires frequent revisions over minor
issues like an extra space in a markdown document, without offering much
in return. An occasional difference in formatting style is fine
as long as it self-corrects in the long run – and that's what the
pre-commit formatting check should take care of.

This is an experimental change and we may have to roll it back or
course-correct if I'm wrong about the self-correcting aspect and the
overall formatting gets wildly out of hand.

That being said, let's give it a try and see how it works.