Prototype: Sync between MEMFS and OPFS
there is one last issue I may not be able to figure out before my
Sabbatical (June 26th - Sep 26th). This PR attempts another approach of
syncing MEMFS changes to OPFS and restoring them after a page refresh.

The main idea is:

1. Compare last modified time
2. Copy MEMFS files to OPFS if they were updated more recently

The current implementation is quite slow as everything happens
synchronously and frequently, but it can be significantly improved.