Add a joinPaths() utility function.
Joining paths in PHP context is commonly needed, however the path.join()
utility is not sufficient. First, it's not available in the browser,
second, it doesn't always produce a POSIX-style paths.

The joinPaths() helper solves this problem in all context where PHP.wasm
is used:

> joinPaths('wordpress', 'wp-content')
'wordpress/wp-content'

Testing instructions:

Confirm the tests are green and the code makes sense