Serve all assets in dev build from same origin
When mounting local directories in #548, problems appeared in development
builds because the dev environemnt runs a separate server for the website
that wraps the iframe and the for the assets inside the iframe, the "remote."

In this patch we're adjusting the config for the outer website server so that
we can proxy requests for the internal frame through the same origin. This is
done by adding a prefix to the URL for the dev assets on the outer frame, which
then is used in a proxy rule to decide what to route.

This means that OPFS handles can be shared between the two sides of the iframe,
which normally happens in production since they both come from the same origin.