Catch PHP initialization errors (#581)
Wraps removeRunDependency to propagate PHP initialization failures. Data
dependencies call removeRunDependency() when they are loaded. The
synchronous call stack then continues to run. If an error occurs in PHP
initialization, e.g. Out Of Memory error, it will not be caught by any
try/catch. This override propagates the failure to PHPLoader.onAbort()
so that it can be handled.

To test, add `throw new Error()` to the `onRuntimeInitialized()`
handler. With this commit, it should trigger Playground error message.