Do not import from wp-playground/remote in wp-playground/website
I'm not sure whether that's a dircular dependency or not, but vite doesn't like
it and starts failing with weird errors like:

> Failed to resolve entry for package "@php-wasm/web". The package may have incorrect main/module/exports specified in its package.json.

My best guess is that vite configuration is different enough between
website and remote to trigger that error. It would be really handy
to have an automated check preventing that import anyway, remote
is not meant to be used as a library.