Add support for WordPress 6.3
Beyond simply regenerating the WordPress files, there was a release bug
in WordPress 6.3 whereby registered styles fail to enqueue in certain
circumstances, one of which exposes itself in the Playground. Adding
`WP_DEVELOPMENT_MODE: all` fixes this bug by bypassing the style
caching mechanism. This has been added to workaround the bug.

This workaround applies if the resolved version of WordPress on
the Playground website is `6.3`, otherwise it skips the step.

See https://core.trac.wordpress.org/ticket/59056