Add fonts directory to file paths (#633)
The Playground resolves URLs based on their path to determine whether to load
the resource at the URL or reroute it to its in-memory storage, such as for
themes and plugin content.

In WordPress 6.4, user-uploaded fonts will be stored in a new top-level uploads
directory and the Playground will need to support this. Otherwise, the Font
paths won't resolve properly and the Playground will return 404 errors for
them.