Accessibility: Font contrast in the URL bar
Form elements are required to have a contrast of at least 4.5, and the
URL bar has a contrast of 4.2 between the background color and the font
color. This commit makes the font lighter to satisfy the requirements.

To test, install axeDevtools and confirm the contrast issue is no longer
reported.

Related: #612