[Patch WordPress] Ensure block editor iframe is controlled by a service worker
Fixes #646

Patches the block editor to use a special ControlledIframe component
instead of a regular HTML "iframe" element. The goal is to make the
iframe use a plain HTTP URL instead of srcDoc, blob URL and other
variations.

Why?

In Playground, the editor iframe needs to be controlled by Playground's service worker
so it can serve CSS and other static assets. Otherwise all the requests originating in
that iframe will yield 404s.

However, different WordPress versions use a variety of iframe techniques
that result in a non-controlled iframe:

* 6.3 uses a binary blob URL and the frame isn't controlled by a service worker
* <= 6.2 uses srcdoc had a null origin and the frame isn't controlled by a service worker
* Other dynamic techniques, such as using a data URL, also fail to
  produce a controlled iframe

HTTP URL src like src="/doc.html" seems to be the only way to create a controlled iframe.

And so, this commit ensures that the block editor iframe uses a plain HTTP
URL regardless of the WordPress version.

Related: https://github.com/WordPress/gutenberg/pull/55152