Browser: Rewrite target="_top" as target="wordpress_playground"
WordPress is rendered in a sandboxed iframe that prevents top-level navigation.
However, some links in wp-admin, such as the "activate plugin" link
rendered after the plugin is uploaded, are rendered with
target="_parent".

It does not make sense to target the top frame in this context. What we really want
is to target the WordPress iframe.

This commit fixes that via a <script> tag that rewrites the `target` of all
clicked A elements from _top and _parent to wordpress-playground.

Related: #266