Fix intermittent JSON schema building errors in CI (#687)
This PR drops the usage of `as const` in `supported-php-extensions.ts`
since they don't seem to play well with the `ts-json-schema-generator`
package that generates Blueprint JSON schema from TypeScript types.

Specifically, it seems to be the root cause behind these intermittent
failures only happening in GitHub actions:

```
UnknownNodeError: Unknown node " supportedPHPExtensions" of kind "Identifier"
```

It's unclear to me why `as const` is fine in
`supported-php-versions.ts`. Perhaps it's because that file uses just
lists and not objects?

I also don't get why the failures are only intermittent. Let's further
investigate if and when these errors return.