Privacy sandbox: Guide the user do when service worker cannot be registered (#694)
The "Allow storage" button does not work with third-party cookies
disabled. This PR adds a useful error message to help the user adjust
their browser settings and run Playground.

This is related to Privacy Sandbox (#586). Embedding Playground on
another domain via an <iframe> may result in a "privacy sandboxed"
environment where Service workers are not available by default.

The behavior subtly differs depending on browser settings:

* Third-party cookies enabled
* Third-party cookies disabled

In each of these modes, the behavior may differ depending on which
experimental flags are enabled (on chrome://flags):

* #storage-access-api
* #third-party-storage-partitioning
* #permission-storage-access-api (maybe?)

Privacy sandbox is a work in progress and the browser implementation
changes over time in response to developer feedback. In Playground, we
need to stay on top of these changes and react to them.

## Testing instructions

Testing this change is tedious!

1. Build Playground
2. Put it on any https:// TLD (secure context is crucial)
3. Embed it via an iframe on any other https:// TLD
4. Go through different combinations of the above settings (third-party
cookies enabled, disabled, then 0, 1, 2, or 3 experimental flags
enabled)
5. Confirm that each scenario yields a useful error message and that
following the steps makes Playground work

CC @andreylipattsev Am I massively overcomplicating this? Is there any
easier way of staying on top of the Privacy Sandbox rollout and testing
PRs like this one?