Request Handler: Serve static files if they exist in VFS. Don't guess.
Replaces approximate rules such as

```js
if(path.startsWith('/wp-content')) {
    // serve from VFS
} else {
    // serve from remote server
}
```

With actual file existence check.

When a static file is requested from Playground, the service worker
doesn't know whether to serve it from VFS or playground.wordpress.net.
It tries to guess that based on approximate rules such as
`path.startsWith('/wp-content')`. This usually works, but is not
very reliable. For example, when you store an entirely new WordPress instance
in VFS, you would expect all the static assets to be served from VFS,
but that is not what the request handler would do.

This does not play well with previewing WordPress Pull Requests as we're
actually replacing the entire WordPress instance in VFS with an
arbitrary branch.

Closes #109

1. Start Playground `npm run start`
2. Use it for a bit, go to wp-admin, confirm that static assets like CSS
   and JS files are loaded without issues
3. Upload something, confirm that worked too

CC @sejas @danielbachhuber as this is a breaking change in public API and may be
relevant for wp-now. While that sounds scary, everything should "just work" without
major changes required in wp-now, other than perhaps removing usages of
the `isStaticFilePath` option.