Add WordPress core Pull Request previewer
Add wordpress.html, a simple HTML form that enables previewing
any WordPress core Pull Request.

Technically, it's very similar to gutenberg.html – it just creates a
tailored Blueprint and redirects to Playground with that Blueprint
applied.

*Caveats*
This leans on storing WordPress builds as a GitHub artifact in CI. For now only one PR does that – WordPress/wordpress-develop#5481. Once it's merged, the previewer proposed in this PR will work for every wordpress-develop Pull Request!

The downloaded WordPress bundle is ~45MB and takes a while to download without any visible progress information. Let's find a way to optimize it.