Download plugins and themes directly from WordPress.org, not via plugin-proxy.php
WordPress.org now serves the `Access-Control-Allow-*` headers necessary
to perform a direct `fetch()` without going through plugin-proxy.php.
This commit thus removes the usages of plugin-proxy.php.

The plugin proxy is now only used by the Pull Request previewers. As a
follow up, let's consider splitting them to a separate package or even
to the playground-tools repo.

To test, go to http://localhost:5400/website-server/?theme=pendant&plugin=gutenberg
and confirm that both the Pendant theme and the Gutenberg plugins are
installed, and also that the zip files are downloaded directly from wordpress.org