[Web WordPress Build] Dont minify files in the wp-content directory (#736)
Something about the twentytwentyfour theme triggers a fatal error if the
PHP files in `wp-content` directory are minified. Since the bundle size
impact is minimal, I'm fixing the problem by just excluding that
directory from the minification process. In the future, let's
investigate and find the root cause of the problem.