Import code from GitHub repo (#733)
## Description:

This PR adds the ability for WordPress Playground to import code from
GitHub repositories. This allows developers to easily pull in themes,
plugins, and `wp-content` directories from GitHub for testing and
development in the Playground environment:


https://github.com/WordPress/wordpress-playground/assets/205419/e05ceaee-c25d-4173-b865-660d904ea853

The import modal requires the user to explicitly state what are they
importing and from which path in the repo, but it also tries to be smart
about guessing the default selection.

## Follow-up work

* Support a file like .playgroundrc that explicitly defines which paths
inside the repo are themes, plugins, or `wp-content` directories
* Support committing changes back to GitHub
* Support Query API parameters to show a pre-populated import modal (to
enable single-click workflows)

## Open questions

* Currently, the OAuth access token isn't stored anywhere. It lives as a
JS variable and gets wiped out on page reload. Is there a way to store
it in such a way that the embedded WordPress couldn't access it?

Supersedes #730