Add php.defineConstant() public API (#748)
https://github.com/WordPress/wordpress-playground/pull/738 introduced a
change that made all WordPress constants be defined by including a
"virtualized" wp-consts.php file. However, the required
`setPhpIniEntry()` call was sometimes happening after the initial
`php.run()` which made it noop. This broke the networking support merged
in #724.

This commit solves that by exposing a new `php.defineConstant()` API
that handles the file virtualization. I'm not super happy about
virtualizing files so let's revisit that in the future, see #750.