Set emptyOutDir to false in php-wasm/web/vide.config.ts
In CI, Vite occasionally fails with the following error:
ENOTEMPTY: directory not empty, rmdir dist/packages/php-wasm/web/light

That directory is implicitly created by preserve-php-loaders-imports
plugin. Vite attempts to synchronously empty it, and yet some files
are somehow left in there. Perhaps something writes them in parallel?
Maybe due to the --parallel=3 Nx setting used in CI?

For now, let's just abstain from emptying that directory. In the future,
if more problems emerge, we may need to get to the bottom of that.