pre-commit hook: Ensure autoformatted files are added to the current commit
The current pre-commit hook does not stage the formatted files for
committing. As a result, files like blueprint-schema.json flicker
back and forth between their formatted and unformatted state as
different contributors have different committing strategies.

This commit is an attempt to standardize that behavior. Every formatted
file that was already staged for a commit, will now be added to that
very commit. This, in theory, should prevent misformatted files from
ever being committed.