WordPress PR Previewer: Improve error messages, retry if the artifact isn't ready yet
Improves the error messages in the WordPress PR previewer, especially when the
WordPress build artifact is not yet ready.

This is important, because the previewer link [is posted on every wordpress-develop PR](https://github.com/WordPress/wordpress-develop/pull/5738)
before the initial build artifact is ready. Simply displaying a generic
error message would be very confusing to users, so this PR tries to
explain the situation and also retry the preview evey 30 seconds.