GitHub Export: Use target repo and branch in preview links.
This fixes a bug where the preview links would always point to the
upstream repo, even if the changes were pushed to the fork. It also
fixes a bug where the preview links would sometimes point to a branch
with a different name than the one the changes were pushed to.

 ## Testing instructions

1. Open Playground, make some changes
2. Click "Export to GitHub" and complete the flow with a repo you have no
   write access to. Keep this tab open.
3. Click "Preview" in the PR description and verify that it loads Playground
   with your changes. Then close that tab.
3. In the original Playground tab, make some more changes, click Export to GitHub,
   select "Update PR", put in the number, then select "Create PR".
4. Complete the flow
5. Click "Preview" in the PR description and verify that it loads Playground