Bidirectional postMessage relay: add relaying from parent window to WordPress inside Playground (#842)
Adds the ability to relay messages from Playground(or 3rd party embed) ==>
WP.

Before this PR, the ability to relay messages between WP and the site
embedding `remote.html` is one direction from WP=>Playground.

## What problem is it solving?

I needed to be able to send messages down to manage state with things in
WP.

## How is the problem addressed?

I've added an event listener that checks the `message` comes from the
`window.parent` which should be the site embedding `remote.html` and the
proper `type: relay` and `object` stuff is included.

> [!RESOLVED]
> **didn't** check the origin cause I was unsure if there was a secure
way to confirm it was from the intended embed...maybe
`window.parent.location.origin`? Not entirely sure this would do
anything but I'm 100% sure I don't understand the security concerns
here.

## Testing Instructions

> [!NOTE]
> I did manual testing. I can work on correct tests if those are a
thing.

> [!RESOLVED]
> I'm not sure this has been correctly tested by hand. The current issue
is that my source domain(Playground) and the `remotehtml` domain are
identical (`localhost:5400`). In my IRL use case, these would be
different. not sure if or how to test this or if it's needed.