Make the .zip export self-contained: Include the default theme, wp-config, and the sqlite-database-integration plugin (#872)
## Description

Includes the `wp-config.php` file and the default WordPress theme (like
`twentytwentythree`) in the zip file exported via the "Download as .zip"
button.

Without those files, the exported bundle isn't self-contained. It cannot
be hosted, and any the importer needs to provide the missing files on
its own. The theme and the plugins are easy to backfill, but the data
stored in `wp-config.php` is lost.

## How is the problem addressed?

This PR adds a temporary private `selfContained` option to the
`importWpContent`. It is `false` by default to ensure those files are
not exported with GitHub PRs (the export flow relies on the same logic).
The zip download button sets it to `true`.

This is a temporary workaround as we currently don't have any better
tools to deal with this problem. Once the streaming/iterators API ships
in https://github.com/WordPress/wordpress-playground/pull/851, we'll be
able to get rid of this hack and just filter the stream of files.

## Testing Instructions

Unfortunately, this PR ships no unit tests as without #895 there isn't
an easy way to test the `zipWpContent` function. Here's the manual
testing steps to take:

1. Open Playground
2. Make a change in the site content
3. Export Playground into a zip file
4. Confirm that zip file contains the `wp-content.php` file as well as
the `twentytwentyfour` theme and the `sqlite-database-integration`
plugin
5. Refresh the Playground tab and import that zip
6. Confirm it worked and the website is functional and has the content
update from before
7. Export it to GitHub, check the "include zip file" checkbox
8. Confirm the GitHub PR has no `twentytwentyfour` theme, or the
`wp-config.php` file, or the `sqlite-database-integration` plugin.
9. Do the same for the zip bundled with the GitHub PR
10. Import that PR and confirm it imports cleanly