Move the stream-compression package from wp-playground to php-wasm
Stream Compression introduced in #851 has no dependencies on WordPress and
can be used in any JavaScript project. It also makes sense as a dependency
for some `@php-wasm` packages. This commit, therefore, moves it from the
`wp-playground` to the `php-wasm` npm namespace, making it reusable across
the entire project.

In addition, this adds a new `iterateFiles` function to the `@php-wasm/universal`
package, which allows iterating over the files in the PHP filesystem. It
uses the `stream-compression` package, which was some of the motivation for
the move.

 ## Testing instructions

Since the package isn't used anywhere yet, only confirm if the CI checks
pass.