Browser: Move patching the Gutenberg iframe into the service worker (#940)
As discussed in
https://github.com/WordPress/wordpress-playground/issues/42,
the Gutenberg plugin needs to be patched to ensure the editor iframe is
controlled by the service worker.

The previous patching approach was to patch the Gutenberg plugin in the
`install-plugin` step. This worked for the `install-plugin` step, but
not for
the `import-site` step, because the Gutenberg plugin is already
installed when
importing a site. That was fixed in another PR, but it's still not
enough.

Turns out, there's a whole lot of ways to install the Gutenberg plugin:

* Install plugin step
* Import a site
* Install Gutenberg from the plugin directory
* Upload a Gutenberg zip

Since it's too difficult to patch Gutenberg in all these cases, this
commit blanket-patches all the scripts requested over the network whose
names seem to indicate they're related to the Gutenberg plugin.