Support wp-cli in the browser (#957)
Adds `wp-cli` support in the browser!

![CleanShot 2024-01-22 at 15 13
38@2x](https://github.com/WordPress/wordpress-playground/assets/205419/4f0a352f-8e2f-4961-b0c4-a9e88804931e)

This PR brings parts of
https://github.com/WordPress/wordpress-playground/pull/161 by
@swissspidy into `trunk` to enable using `wp-cli` in the browser.

## Codebase changes

* `setSapiName` method for PHP. `wp-cli` expects to see the exact string
`"cli"` as the SAPI name.
*  All PHP versions are now built with phar support.
* `js_open_process` is asynchronous via Asyncify. This is to enable
calling `setSpawnHandler(handler)` from the parent site to set a handler
in an asynchronous web worker. Unfortunately, using the Comlink
messaging library is challenging with complex objects like spawn
handlers so this PR ships a workaround where spawn handlers are passed
as strings and `eval()`-ed in the worker. This shouldn't have any
security implications as Playground enables evaluating arbitrary code by
definition.
* `proc_open` waits for the process to spawn, and `proc_close` waits for
the process to die. This is necessary to support the `wp-cli` behavior
of piping its output through a pager program (like `cat` or `less -S`).

## Follow-up work

* Remove `eval()` from `setSpawnHandler(handler)`. Setting spawn handler
works in Node.js where PHP runs synchronously in the same process, but
these spawn handlers are challenging to transfer over `postMessage` as
they're essentially composite EventEmitters that need to accept input
inside the Worker process, handle it in the client process, and transmit
it back to the worker process. Intuitively, it would be pretty easy if
we used `MessagePort` directly. It seems like `Comlink` as an
abstraction layer makes things much easier to a point, since we can
magically call functions on objects living in another thread, but
there's a boundary beyond which Comlink makes these API interactions
more difficult.


cc @danielbachhuber @schlessera