Blueprints: Add enableMultisite step
This step enables multisite support in the WordPress installation. It
does not create any sites, but it does create the necessary tables in
the database.

Example Blueprint:

```json
{
    "landingPage":"/wp-admin/network.php",
    "steps": [
        { "step": "enableMultisite" }
    ]
}
```

 ## URL rewriting

This PR also ships a URL rewriting logic in the service worker. This
is to ensure that requests to URLs like `/mysite/wp-admin/` are still
resolved to the `/wp-admin/index.php` file.

Internally, this is done by adding a `x-rewrite-url` header to the
request whenever the URL matches a rewrite rule after an initial 404.

CC @dmsnell