Use .zip files instead of .data files for loading WordPress (#978)
Playground used the Emscripten's filepacker.py tool to bundle WordPress
releases as zip files. This caused a lot of problems:

* Filepacker splits WordPress into a `.data` file, with all the bytes
concatenated as a single blob, and a `.js` file with the map of offsets
and lengths for each file. Sometimes the browsers would invalidate the
cache for the `.data` file, but not the `.js` file, causing the
playground to fail to load.
* The `.js` files had to be patched in the Dockerfile.
* The `.js` files used XHR to load the `.data` files, which required
custom error handling logic, progress monitoring logic, and even then
some of the errors were not handled properly – see #564.
* The `.data` loading code required custom loading logic.
* The `.data` files were not stream–loaded. This PR doesn't stream-load
the `.zip` files either, but that feature may be added in the future
using the new `@wp-playground/compression-streams` package.

## Changes

* The WordPress build process now produces `.zip` files instead of
`.data`
  + `.js` files.
* The worker thread now loads WordPress from zip files, including zips
from arbitrary URLs.
* Query API and Blueprints now accept URLs via the `?wp=` parameter.

## Testing instructions

* Confirm all the E2E tests pass

cc @bgrgicak